requirements.txt
pytest==6.2.5
django==3.2
python-dotenv==0.19.2
django-split-settings==1.1.0